backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt
Pillow==10.3.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
sqlparse==0.5.5
tzdata==2025.3
